Fix compiling

+++ core/src/main/java/de/gerrygames/viarewind/protocol/protocol1_7_6_10to1_8/packets/EntityPackets.java
@@ -79,7 +79,7 @@ public class EntityPackets {
 					public void handle(PacketWrapper packetWrapper) throws Exception {
 						Position position = packetWrapper.read(Type.POSITION);
 						packetWrapper.write(Type.INT, position.getX());
-						packetWrapper.write(Type.UNSIGNED_BYTE, position.getY());
+						packetWrapper.write(Type.UNSIGNED_BYTE, (short) position.getY());
 						packetWrapper.write(Type.INT, position.getZ());
 					}
 				});
